add new --kubernetes-image-pull-secrets parameter

+++ pkg/executors/kubernetes_executor.go
@@ -57,7 +57,6 @@ func NewKubernetesExecutor(jobRequest *api.JobRequest, logger *eventlogger.Logge
 func (e *KubernetesExecutor) Prepare() int {
 	e.podName = e.randomPodName()
 	e.envSecretName = fmt.Sprintf("%s-secret", e.podName)
-	e.imagePullSecret = fmt.Sprintf("%s-image-pull-secret", e.podName)
 
 	err := e.k8sClient.CreateSecret(e.envSecretName, e.jobRequest)
 	if err != nil {
@@ -65,10 +64,15 @@ func (e *KubernetesExecutor) Prepare() int {
 		return 1
 	}
 
-	err = e.k8sClient.CreateImagePullSecret(e.imagePullSecret, e.jobRequest.Compose.ImagePullCredentials)
-	if err != nil {
-		log.Errorf("Error creating image pull credentials '%s': %v", e.envSecretName, err)
-		return 1
+	// If image pull credentials are specified in the YAML,
+	// we create a temporary secret to store them and use it to pull the image.
+	if len(e.jobRequest.Compose.ImagePullCredentials) > 0 {
+		e.imagePullSecret = fmt.Sprintf("%s-image-pull-secret", e.podName)
+		err = e.k8sClient.CreateImagePullSecret(e.imagePullSecret, e.jobRequest.Compose.ImagePullCredentials)
+		if err != nil {
+			log.Errorf("Error creating image pull credentials '%s': %v", e.envSecretName, err)
+			return 1
+		}
 	}
 
 	err = e.k8sClient.CreatePod(e.podName, e.envSecretName, e.imagePullSecret, e.jobRequest)
@@ -359,9 +363,14 @@ func (e *KubernetesExecutor) removeK8sResources() {
 		log.Errorf("Error deleting secret '%s': %v\n", e.envSecretName, err)
 	}
 
-	err = e.k8sClient.DeleteSecret(e.imagePullSecret)
-	if err != nil {
-		log.Errorf("Error deleting secret '%s': %v\n", e.imagePullSecret, err)
+	// Not all jobs create this temporary secret,
+	// just the ones that send credentials to pull images
+	// in the job definition, so we only delete it if it was previously created.
+	if e.imagePullSecret != "" {
+		err = e.k8sClient.DeleteSecret(e.imagePullSecret)
+		if err != nil {
+			log.Errorf("Error deleting secret '%s': %v\n", e.imagePullSecret, err)
+		}
 	}
 }
